Complete custom hooks refactor
